+++ src/views/hostnetworkconfiguration/components/BridgeIcon.tsx
