+++ src/views/hostnetworkconfiguration/components/CustomGroup/CustomGroup.scss
@@ -1,0 +1,10 @@
+.custom-group {
+  .pf-topology__group__label > text {
+    fill: var(--pf-topology__node__label__text--Fill);
+  }
+
+  .pf-topology__node__label__background {
+    fill: var(--pf-topology__node__label__background--Fill);
+    stroke: var(--pf-topology__node__background--Stroke);
+  }
+}
