+++ src/views/hostnetworkconfiguration/components/TopologySidebar/TopologySidebar.scss
