+++ src/views/hostnetworkconfiguration/utils/constants.ts
