+++ src/views/hostnetworkconfiguration/utils/factory.ts
