+++ src/views/hostnetworkconfiguration/utils/utils.ts
