+++ src/views/states/list/components/InterfaceDrawer/InterfaceDrawer.tsx
@@ -1,15 +1,16 @@
 import React, { FC } from 'react';
+import { useLocation } from 'react-router-dom-v5-compat';
 
 import {
-  PageSection,
-  Title,
-  Tabs as TabsComponent,
-  Tab,
-  TabTitleText,
   Modal,
   ModalBody,
   ModalFooter,
   ModalHeader,
+  PageSection,
+  Tab,
+  Tabs as TabsComponent,
+  TabTitleText,
+  Title,
 } from '@patternfly/react-core';
 import { NodeNetworkConfigurationInterface } from '@types';
 import { useNMStateTranslation } from '@utils/hooks/useNMStateTranslation';
@@ -26,6 +27,8 @@ type InterfaceDrawerProps = {
 
 const InterfaceDrawer: FC<InterfaceDrawerProps> = ({ selectedInterface, onClose }) => {
   const { t } = useNMStateTranslation();
+
+  const location = useLocation();
 
   const Tabs: InterfaceDrawerTabProps[] = [
     {
